feat：完成新增罚款记录接口

+++ src/main/java/com/mumu/emos/api/db/pojo/Amect.java
@@ -36,7 +36,7 @@ public class Amect {
     /**
      * 罚款类型
      */
-    private Integer typeId;
+    private Byte typeId;
 
     /**
      * 罚款原因
